Moved files around like a game of musical chairs. Nobody's left standing!
