+++ tests/unit/Agents/GP/CGPGenotype.cpp
@@ -1,13 +1,13 @@
 /**
  * This file is part of the Fall 2023, CSE 491 course project.
- * @brief Unit tests for source/Group7_GP_Agent/CGPGenotype.hpp
+ * @brief Unit tests for source/[Deprecated]Group7_GP_Agent/CGPGenotype.hpp
  **/
 
 // Catch2
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch_all.hpp>
 
-#include "Group7_GP_Agent/CGPGenotype.hpp"
+#include "Agents/GP/CGPGenotype.hpp"
 #include <ranges>
 
 using namespace cowboys;
